Refactors and introduction of test cases for the pipeline munge.
